refactor: move pause prompt and styledInput to executor_input_helpers.go
